back/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
alembic
psycopg2-binary
python-jose[cryptography]
bcrypt
pydantic
pytest
httpx 